shuffle: move all shufflers to namespace
