+++ BlasII.Randomizer/Shuffle/ReverseItemShuffler.cs
@@ -1,22 +1,23 @@
 ﻿using System;
 using System.Collections.Generic;
+using BlasII.Randomizer.Items;
 
-namespace BlasII.Randomizer.Items.Shuffle
+namespace BlasII.Randomizer.Shuffle
 {
-    internal class PoolsItemShuffler : IShuffler
+    internal class ReverseItemShuffler : IShuffler
     {
         public bool Shuffle(int seed, RandomizerSettings settings, Dictionary<string, string> output)
         {
             output.Clear();
-            var rng = new Random(seed);
+            Initialize(seed);
 
             // Create pools of all locations to randomize
-            LocationPool progressionLocations = new(rng), junkLocations = new(rng);
+            List<ItemLocation> progressionLocations = new(), junkLocations = new();
             CreateLocationPool(progressionLocations, junkLocations, settings);
 
             // Create pools of all items to randomize
-            ItemPool progressionItems = new(rng), junkItems = new(rng);
-            CreateItemPool(progressionItems, junkItems, progressionLocations.Size + junkLocations.Size, settings);
+            List<Item> progressionItems = new(), junkItems = new();
+            CreateItemPool(progressionItems, junkItems, progressionLocations.Count + junkLocations.Count, settings);
 
             // Create initial inventory
             var inventory = new Blas2Inventory(settings, Main.Randomizer.Data.DoorDictionary);
@@ -26,15 +27,15 @@ namespace BlasII.Randomizer.Items.Shuffle
             FillProgressionItems(progressionLocations, progressionItems, output, inventory);
 
             // Add junk items to remaining progression items and ensure they are still balanced
-            junkLocations.Combine(progressionLocations);
-            if (junkLocations.Size != junkItems.Size)
+            junkLocations.AddRange(progressionLocations);
+            if (junkLocations.Count != junkItems.Count)
                 return false;
 
             // Place junk items at junk locations and remaining progression locations
             FillJunkItems(junkLocations, junkItems, output);
 
             // Verify that all remaining items were placed
-            return junkItems.Size == 0 && junkLocations.Size == 0;
+            return junkItems.Count == 0 && junkLocations.Count == 0;
         }
 
         #region Setup
@@ -42,7 +43,7 @@ namespace BlasII.Randomizer.Items.Shuffle
         /// <summary>
         /// Fills the two location pools
         /// </summary>
-        private void CreateLocationPool(LocationPool progressionLocations, LocationPool junkLocations, RandomizerSettings settings)
+        private void CreateLocationPool(List<ItemLocation> progressionLocations, List<ItemLocation> junkLocations, RandomizerSettings settings)
         {
             foreach (var location in Main.Randomizer.Data.ItemLocationList)
             {
@@ -53,16 +54,16 @@ namespace BlasII.Randomizer.Items.Shuffle
         /// <summary>
         /// Takes a single location data and adds it to the correct list based on its type
         /// </summary>
-        private void AddLocationToPool(LocationPool progressionLocations, LocationPool junkLocations, ItemLocation location, RandomizerSettings settings)
+        private void AddLocationToPool(List<ItemLocation> progressionLocations, List<ItemLocation> junkLocations, ItemLocation location, RandomizerSettings settings)
         {
-            LocationPool locationPool = location.ShouldBeShuffled(settings) ? progressionLocations : junkLocations;
+            List<ItemLocation> locationPool = location.ShouldBeShuffled(settings) ? progressionLocations : junkLocations;
             locationPool.Add(location);
         }
 
         /// <summary>
         /// Fills the two item pools to match the number of locations
         /// </summary>
-        private void CreateItemPool(ItemPool progressionItems, ItemPool junkItems, int numOfLocations, RandomizerSettings settings)
+        private void CreateItemPool(List<Item> progressionItems, List<Item> junkItems, int numOfLocations, RandomizerSettings settings)
         {
             foreach (var item in Main.Randomizer.Data.ItemList)
             {
@@ -76,9 +77,9 @@ namespace BlasII.Randomizer.Items.Shuffle
         /// <summary>
         /// Takes a single item data and adds it to the correct list based on its type
         /// </summary>
-        private void AddItemToPool(ItemPool progressionItems, ItemPool junkItems, Item item)
+        private void AddItemToPool(List<Item> progressionItems, List<Item> junkItems, Item item)
         {
-            ItemPool itemPool = item.progression ? progressionItems : junkItems;
+            List<Item> itemPool = item.progression ? progressionItems : junkItems;
             for (int i = 0; i < item.count; i++)
             {
                 itemPool.Add(item);
@@ -88,7 +89,7 @@ namespace BlasII.Randomizer.Items.Shuffle
         /// <summary>
         /// Removes the starting weapon from the item pool
         /// </summary>
-        private void RemoveStartingItemsFromItemPool(ItemPool items, RandomizerSettings settings)
+        private void RemoveStartingItemsFromItemPool(List<Item> items, RandomizerSettings settings)
         {
             // Remove the extra starting weapon
             items.Remove(Main.Randomizer.Data.GetItem(GetStartingWeaponId(settings)));
@@ -97,16 +98,16 @@ namespace BlasII.Randomizer.Items.Shuffle
         /// <summary>
         /// After creating the item pool, add or remove junk items to make it equal to the number of locations
         /// </summary>
-        private void BalanceItemPool(ItemPool progressionItems, ItemPool junkItems, int numOfLocations)
+        private void BalanceItemPool(List<Item> progressionItems, List<Item> junkItems, int numOfLocations)
         {
             // Remove tear items until pools are equal
-            while (progressionItems.Size + junkItems.Size > numOfLocations)
+            while (progressionItems.Count + junkItems.Count > numOfLocations)
             {
-                junkItems.RemoveLast();
+                junkItems.RemoveAt(junkItems.Count - 1);
             }
 
             // Add tear items until pools are equal
-            while (progressionItems.Size + junkItems.Size < numOfLocations)
+            while (progressionItems.Count + junkItems.Count < numOfLocations)
             {
                 junkItems.Add(Main.Randomizer.Data.GetItem("Tears[800]"));
             }
@@ -115,7 +116,7 @@ namespace BlasII.Randomizer.Items.Shuffle
         /// <summary>
         /// Adds the starting weapon and all progression items to the initial inventory
         /// </summary>
-        private void CreateInitialInventory(Blas2Inventory inventory, RandomizerSettings settings, ItemPool progressionItems)
+        private void CreateInitialInventory(Blas2Inventory inventory, RandomizerSettings settings, List<Item> progressionItems)
         {
             // Add the starting weapon
             inventory.AddItem(Main.Randomizer.Data.GetItem(GetStartingWeaponId(settings)));
@@ -142,7 +143,7 @@ namespace BlasII.Randomizer.Items.Shuffle
         /// <summary>
         /// Using a reverse fill algorithm, place the last item at a random reachable location
         /// </summary>
-        private void FillProgressionItems(LocationPool locations, ItemPool items, Dictionary<string, string> output, Blas2Inventory inventory)
+        private void FillProgressionItems(List<ItemLocation> locations, List<Item> items, Dictionary<string, string> output, Blas2Inventory inventory)
         {
             // Verify that all locations are reachable
             foreach (var location in locations)
@@ -151,20 +152,20 @@ namespace BlasII.Randomizer.Items.Shuffle
                     return;
             }
 
-            items.Shuffle();
+            ShuffleList(items);
             MovePriorityItems(items);
-            var reachableLocations = new LocationPool(locations);
+            var reachableLocations = new List<ItemLocation>(locations);
 
-            while (reachableLocations.Size > 0 && items.Size > 0)
+            while (reachableLocations.Count > 0 && items.Count > 0)
             {
-                Item item = items.RemoveLast();
+                Item item = RemoveLast(items);
                 inventory.RemoveItem(item);
 
                 RemoveUnreachableLocations(reachableLocations, inventory);
-                if (reachableLocations.Size == 0)
+                if (reachableLocations.Count == 0)
                     return;
 
-                ItemLocation location = reachableLocations.RemoveRandom();
+                ItemLocation location = RemoveRandom(reachableLocations);
                 locations.Remove(location);
                 output.Add(location.id, item.id);
             }
@@ -173,14 +174,14 @@ namespace BlasII.Randomizer.Items.Shuffle
         /// <summary>
         /// Without logic, place the last item at a random location
         /// </summary>
-        private void FillJunkItems(LocationPool locations, ItemPool items, Dictionary<string, string> output)
+        private void FillJunkItems(List<ItemLocation> locations, List<Item> items, Dictionary<string, string> output)
         {
-            items.Shuffle();
+            ShuffleList(items);
 
-            while (locations.Size > 0 && items.Size > 0)
+            while (locations.Count > 0 && items.Count > 0)
             {
-                ItemLocation location = locations.RemoveLast();
-                Item item = items.RemoveLast();
+                ItemLocation location = RemoveLast(locations);
+                Item item = RemoveLast(items);
 
                 output.Add(location.id, item.id);
             }
@@ -189,20 +190,77 @@ namespace BlasII.Randomizer.Items.Shuffle
         /// <summary>
         /// After shuffling the list of progression items, move certain items to the end to prevent failing seeds
         /// </summary>        
-        private void MovePriorityItems(ItemPool progressionItems)
+        private void MovePriorityItems(List<Item> progressionItems)
         {
             Item wallClimb = Main.Randomizer.Data.GetItem("AB44");
-            progressionItems.MoveToBeginning(wallClimb);
+            progressionItems.Remove(wallClimb);
+            progressionItems.Insert(0, wallClimb);
         }
 
         /// <summary>
         /// Finds all locations that are no longer reachable and removes them
         /// </summary>
-        private void RemoveUnreachableLocations(LocationPool locations, Blas2Inventory inventory)
+        private void RemoveUnreachableLocations(List<ItemLocation> locations, Blas2Inventory inventory)
         {
-            locations.RemoveConditional(loc => !inventory.Evaluate(loc.logic));
+            for (int i = 0; i < locations.Count; i++)
+            {
+                if (inventory.Evaluate(locations[i].logic))
+                    continue;
+
+                locations.RemoveAt(i);
+                i--;
+            }
         }
 
         #endregion
+
+        // Old base shuffler stuff
+
+        private Random rng;
+
+        protected void Initialize(int seed) => rng = new Random(seed);
+
+        protected int RandomInteger(int max) => rng.Next(max);
+
+        protected T RandomElement<T>(List<T> list) => list[RandomInteger(list.Count)];
+
+        protected void ShuffleList<T>(List<T> list)
+        {
+            int upperIdx = list.Count;
+            while (upperIdx > 1)
+            {
+                upperIdx--;
+                int randIdx = RandomInteger(upperIdx + 1);
+                T value = list[randIdx];
+                list[randIdx] = list[upperIdx];
+                list[upperIdx] = value;
+            }
+        }
+
+        protected T RemoveRandom<T>(List<T> list)
+        {
+            int index = RandomInteger(list.Count);
+            T element = list[index];
+
+            list.RemoveAt(index);
+            return element;
+        }
+
+        protected T RemoveRandomFromOther<T>(List<T> list, List<T> other)
+        {
+            T element = RandomElement(list);
+
+            other.Remove(element);
+            return element;
+        }
+
+        protected T RemoveLast<T>(List<T> list)
+        {
+            int index = list.Count - 1;
+            T element = list[index];
+
+            list.RemoveAt(index);
+            return element;
+        }
     }
 }
